:recycle: chore(api) extract service from model
